+++ packages/hierarchy-menu/src/helperComponents/TreeMenu/TreeMenu.tsx
@@ -1,19 +1,26 @@
 import { Divider } from '@sbercloud/uikit-react-divider';
+import { FC } from 'react';
 
 import { TreeNode, TreeNodeProps } from '../TreeNode';
 import * as S from './styled';
 
-export function TreeMenu({
-  activeKey,
-  setActiveKey,
-  ...treeNodeProps
-}: TreeNodeProps & { setActiveKey(x: string): void; activeKey: string }) {
-  return (
-    <>
-      <TreeNode {...treeNodeProps} key={treeNodeProps.id} setActiveKey={setActiveKey} activeKey={activeKey} />
-      <S.DividerWrapper>
-        <Divider />
-      </S.DividerWrapper>
-    </>
-  );
+interface TreeMenu {
+  setActiveKey(x: string): void;
+  activeKey: string;
+  onItemClick?(id: string): void;
 }
+
+export const TreeMenu: FC<TreeMenu & TreeNodeProps> = ({ activeKey, setActiveKey, onItemClick, ...treeNodeProps }) => (
+  <>
+    <TreeNode
+      {...treeNodeProps}
+      key={treeNodeProps.id}
+      setActiveKey={setActiveKey}
+      activeKey={activeKey}
+      onItemClick={onItemClick}
+    />
+    <S.DividerWrapper>
+      <Divider />
+    </S.DividerWrapper>
+  </>
+);