+++ packages/modal/src/components/Default/Modal.tsx
@@ -1,7 +1,7 @@
-import { CloseInterfaceSVG } from '@sbercloud/uikit-react-icons';
 import { Button, IconButton } from '@sbercloud/uikit-react-button';
-import { useLanguage } from '@sbercloud/uikit-utils';
+import { CloseInterfaceSVG } from '@sbercloud/uikit-react-icons';
 import { Tooltip } from '@sbercloud/uikit-react-tooltip';
+import { useLanguage } from '@sbercloud/uikit-utils';
 import { useMemo } from 'react';
 import RCModal from 'react-modal';
 
@@ -63,6 +63,7 @@ export interface ModalProps extends ReactModalProps {
   contentStyles?: React.CSSProperties;
   zIndex?: number;
   parentId?: string;
+  alarmApproveButton?: boolean;
 }
 
 export const MODAL_CLOSE_TYPE = {
@@ -90,6 +91,7 @@ export const Modal: React.FC<ModalProps> = props => {
     zIndex = 99999,
     parentId,
     parentSelector,
+    alarmApproveButton,
   } = props;
 
   const { code: language } = useLanguage({ onlyEnabledLanguage: true });
@@ -137,6 +139,7 @@ export const Modal: React.FC<ModalProps> = props => {
                 <Button
                   className={buttonCSS}
                   disabled={disableApprove}
+                  variant={alarmApproveButton ? Button.variants.Alarm : Button.variants.Filled}
                   onClick={(e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
                     approve(e);
                     onRequestClose?.(e, MODAL_CLOSE_TYPE.APPROVE);
@@ -147,6 +150,7 @@ export const Modal: React.FC<ModalProps> = props => {
               </Tooltip>
             ) : (
               <Button
+                variant={alarmApproveButton ? Button.variants.Alarm : Button.variants.Filled}
                 className={buttonCSS}
                 disabled={disableApprove}
                 onClick={(e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {