+++ packages/radio/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.4](https://bitbucket.sbercloud.tech/projects/SBERCLOUD_UI/repos/uikit2.0/compare/diff?targetBranch=refs/tags/@sbercloud/uikit-react-radio@0.3.3&sourceBranch=refs/tags/@sbercloud/uikit-react-radio@0.3.4) (2021-07-21)
+
+**Note:** Version bump only for package @sbercloud/uikit-react-radio
+
+
+
+
+
+## [0.3.3](https://bitbucket.sbercloud.tech/projects/SBERCLOUD_UI/repos/uikit2.0/compare/diff?targetBranch=refs/tags/@sbercloud/uikit-react-radio@0.3.2&sourceBranch=refs/tags/@sbercloud/uikit-react-radio@0.3.3) (2021-07-20)
+
+**Note:** Version bump only for package @sbercloud/uikit-react-radio
+
+
+
+
+
 ## [0.3.2](https://bitbucket.sbercloud.tech/projects/SBERCLOUD_UI/repos/uikit2.0/compare/diff?targetBranch=refs/tags/@sbercloud/uikit-react-radio@0.3.1&sourceBranch=refs/tags/@sbercloud/uikit-react-radio@0.3.2) (2021-07-16)
 
 **Note:** Version bump only for package @sbercloud/uikit-react-radio