+++ packages/select/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.11.5](https://bitbucket.sbercloud.tech/projects/SBERCLOUD_UI/repos/uikit2.0/compare/diff?targetBranch=refs/tags/@sbercloud/uikit-react-select@0.11.4&sourceBranch=refs/tags/@sbercloud/uikit-react-select@0.11.5) (2021-07-21)
+
+**Note:** Version bump only for package @sbercloud/uikit-react-select
+
+
+
+
+
+## [0.11.4](https://bitbucket.sbercloud.tech/projects/SBERCLOUD_UI/repos/uikit2.0/compare/diff?targetBranch=refs/tags/@sbercloud/uikit-react-select@0.11.3&sourceBranch=refs/tags/@sbercloud/uikit-react-select@0.11.4) (2021-07-20)
+
+**Note:** Version bump only for package @sbercloud/uikit-react-select
+
+
+
+
+
 ## [0.11.3](https://bitbucket.sbercloud.tech/projects/SBERCLOUD_UI/repos/uikit2.0/compare/diff?targetBranch=refs/tags/@sbercloud/uikit-react-select@0.11.2&sourceBranch=refs/tags/@sbercloud/uikit-react-select@0.11.3) (2021-07-16)
 
 **Note:** Version bump only for package @sbercloud/uikit-react-select